update type to f64

+++ src/model.rs
@@ -1,7 +1,7 @@
 // making a basic model 
 // Initially only taking a single parameter w
 
-pub fn prediction(x : i64) -> i64 {
-    let w = 1;
+pub fn prediction(x : f64) -> f64 {
+    let w = 1.0;
     return w * x;
 }
